[FTR] Rebalance alerting tests
