+++ x-pack/platform/test/alerting_api_integration/security_and_spaces/group4/tests/alerting/event_log.ts
@@ -12,7 +12,7 @@ import { descriptorToArray } from '@kbn/encrypted-saved-objects-plugin/server/cr
 import { Spaces } from '../../../scenarios';
 import { getUrlPrefix, getTestRuleData, ObjectRemover, getEventLog } from '../../../../common/lib';
 import type { FtrProviderContext } from '../../../../common/ftr_provider_context';
-import { validateEvent } from '../../../../spaces_only/tests/alerting/group1/event_log';
+import { validateEvent } from '../../../../spaces_only/tests/alerting/validate_event';
 
 export default function eventLogTests({ getService }: FtrProviderContext) {
   const supertest = getService('supertest');
