+++ x-pack/platform/test/alerting_api_integration/spaces_only/tests/alerting/group1/index.ts
@@ -28,7 +28,5 @@ export default function alertingTests({ loadTestFile, getService }: FtrProviderC
     loadTestFile(require.resolve('./get_rule_tags'));
     loadTestFile(require.resolve('./rule_types'));
     loadTestFile(require.resolve('./rule_types_internal'));
-    // the next file takes ~37 min to run
-    loadTestFile(require.resolve('./event_log'));
   });
 }
