+++ x-pack/platform/test/alerting_api_integration/spaces_only/tests/alerting/group5/index.ts
@@ -1,0 +1,18 @@
+/*
+ * Copyright Elasticsearch B.V. and/or licensed to Elasticsearch B.V. under one
+ * or more contributor license agreements. Licensed under the Elastic License
+ * 2.0; you may not use this file except in compliance with the Elastic License
+ * 2.0.
+ */
+
+import type { FtrProviderContext } from '../../../../common/ftr_provider_context';
+import { buildUp, tearDown } from '../../helpers';
+
+export default function alertingTests({ loadTestFile, getService }: FtrProviderContext) {
+  describe('Alerting', () => {
+    before(async () => await buildUp(getService));
+    after(async () => await tearDown(getService));
+    // the next file takes ~37 min to run
+    loadTestFile(require.resolve('./event_log'));
+  });
+}
